requirements.txt
numpy
rx
sounddevice
websocket-client